pyproject.toml
[tool.poetry]
name = "version_exporter"
version = "0.1.0"
description = "Prometheus exporter for RPM packages"
authors = ["Charles Bourget <charles.bourget.2@ens.etsmtl.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.6"
PyYAML = "^6.0"
prometheus-client = "^0.13.1"
Cerberus = "^1.3.4"

[tool.poetry.dev-dependencies]
poethepoet = "^0.13.1"
pylint = "^2.12.2"
black = "^22.1.0"
mypy = "^0.941"
types-PyYAML = "^6.0.5"
rstcheck = "^3.3.1"
Pygments = "^2.11.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
  [tool.poe.tasks.format]
  help = "Run black on the code base"
  cmd  = "black ."
  [tool.poe.tasks.types]
  help = "Run the type checker"
  cmd  = "mypy . --ignore-missing-imports"
  [tool.poe.tasks.lint]
  help = "Run pylint"
  cmd = "pylint version_exporter"
  [tool.poe.tasks.check-docs]
  help = "Validate rst syntax in the docs"
  cmd  = "rstcheck README.rst"
  [tool.poe.tasks.check]
  help     = "Run all checks on the code base"
  sequence = ["check-docs", "format", "types", "lint"]